\name{mvLSWimpute-package}
\alias{mvLSWimpute-package}
\alias{mvLSWimpute}
\docType{package}
\title{
\packageTitle{mvLSWimpute}
}
\description{
\packageDescription{mvLSWimpute}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mvLSWimpute}
\packageIndices{mvLSWimpute}

The main routine of the package is \code{\link{mv_impute}} which performs forward or forward and backward imputation of locally stationary multivariate time series, using one-step ahead forecasting (and backcasting).  
}
\author{
\packageAuthor{mvLSWimpute}

Maintainer: \packageMaintainer{mvLSWimpute}
}
\references{
Wilson, R. E., Eckley, I. A., Nunes, M. A. and Park, T. (2021) A wavelet-based approach for imputation in nonstationary multivariate time series.
_Statistics and Computing_ *31* Article 18, doi:10.1007/s11222-021-09998-2.
}

\keyword{ package }
